Update BankAcc.java
Made the class abstract, since nobody should instantiate this bankaccount, and a lot of functions are account dependent. So, they'll
be overriden anyway.

I changed the constructor to only include a financial tracker, since i thought that the customer will have a record of the bankaccounts. So, we cant have the bank reference the customer and the customer reference the bank.